requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
aiohttp==3.9.3
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2024.1
numpy==1.26.3
pandas==2.2.0
scipy>=1.7.0
statsmodels>=0.13.0
scikit-learn>=1.0.0
psutil>=5.9.0